Camera ray
Somehow managed to successfully implement a camera picker but its ONLY
from RAW mouse movement by having the center of the screen be 0,0 but i
would need to change this for the actual worldbuilder if i don't want
raw mouse movement. That said, i could still do this but by freeing
mouse to click on icons.

Lighting

I don't think lighting really worked and im still trying to find out why
but that seems to be the best i can get so far.