GUI UPDATE
Got slider working and ability to go between free mouse and normal
mouse, horray!
Now, next step for GUIs is to massively clean everything up because the
code to get this working is very clunky, messy and just poorly
implemented and unflexible and hacky. This is the next on the to do