Terrain Editing
I have finally been able to create some sort of system to allow me to
edit the terrain BUT the quality of the terrain (i.e how smooth the mesh
is) is horrible, if the point where you try to move is not EXACTLY at a
vertex, that point's height won't be reflected by the terrain. Also, the
smoothness of a curve on the terrain is also quite questionable because
of how jagged it is, so for example if the point is EXACTLY at a vertex,
all the other points would be so much lower down from the vertex that it
becomes very pointy.

Additionally, because of the way I've implemented my terrain, it has to
send ALL the VAO attribute data instead of just the vertex data. I may
need to create a new class just for the terrain in order to do this or
I'll have to refactor a lot of my code.